requirements.txt
altair==5.5.0
asgiref==3.10.0
astroid==4.0.1
attrs==24.3.0
awscli==1.36.38
blinker==1.9.0
botocore==1.35.97
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
Django==5.2.7
docopt==0.6.2
docutils==0.16
enum-compat==0.0.3
filelock==3.16.1
flake8==7.3.0
Flask==2.3.3
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
huggingface-hub==0.27.0
idna==3.10
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
llvmlite==0.45.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.20.1
networkx==3.5
numba==0.62.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pipreqs==0.4.13
platformdirs==4.5.0
protobuf==5.29.2
psutil==7.1.0
pyarrow==18.1.0
pyasn1==0.6.1
pycodestyle==2.14.0
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.0
pylint==4.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.7.2
s3transfer==0.10.4
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
setuptools==75.7.0
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.41.1
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torch-model-archiver==0.12.0
torchserve==0.12.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
xgboost==3.0.5
yarg==0.1.10
